umunthu-ai/requirements..txt
Flask
openai
pinecone-client
langchain-core
langchain-openai
langchain-pinecone
python-dotenv
numpy
pandas
